move enums to `Enums`
